+++ unsafe_collection/src/bytes/buf_list.rs
@@ -53,7 +53,10 @@ impl<B: Buf, const LEN: usize> BufList<B, LEN> {
     }
 }
 
-impl<B: Buf, const LEN: usize> Buf for BufList<B, LEN> {
+impl<B, const LEN: usize> Buf for BufList<B, LEN>
+where
+    B: Buf,
+{
     #[inline]
     fn remaining(&self) -> usize {
         self.remaining
@@ -129,6 +132,28 @@ impl<B: Buf, const LEN: usize> Buf for BufList<B, LEN> {
                 bm.freeze()
             }
         }
+    }
+}
+
+impl<B, const LEN: usize> From<BufList<B, LEN>> for BytesMut
+where
+    B: Buf,
+{
+    fn from(mut buf: BufList<B, LEN>) -> Self {
+        let mut bytes = BytesMut::with_capacity(buf.remaining);
+        while let Some(buf) = buf.bufs.pop_front() {
+            bytes.extend_from_slice(buf.chunk());
+        }
+        bytes
+    }
+}
+
+impl<B, const LEN: usize> From<BufList<B, LEN>> for Bytes
+where
+    B: Buf,
+{
+    fn from(buf: BufList<B, LEN>) -> Self {
+        BytesMut::from(buf).into()
     }
 }